+++ options/Credits.d.ts
@@ -5,8 +5,8 @@
  * A valid license is required for using this software.
  * See highcharts.com/license
  *
- * Built for Highcharts v12.6.0.
- * Build stamp: 2026-06-10
+ * Built for Highcharts v13.0.0.
+ * Build stamp: 2026-06-30
  *
  */
 export type CreditsProps = {
@@ -18,6 +18,7 @@ export type CreditsProps = {
     position?: Highcharts.AlignObject;
     style?: Highcharts.CSSObject;
     text?: string;
+    useHTML?: boolean;
     children?: string | (string | number)[] | React.ReactElement | React.ReactElement[];
 };
 export declare function Credits(props: CreditsProps): any;