+++ options/DataTable.d.ts
@@ -1,0 +1,26 @@
+/**
+ * React integration.
+ * Copyright (c) 2026, Highsoft
+ *
+ * A valid license is required for using this software.
+ * See highcharts.com/license
+ *
+ * Built for Highcharts v13.0.0.
+ * Build stamp: 2026-06-30
+ *
+ */
+export type DataTableProps = {
+    columns?: Highcharts.DataTableColumnCollection | undefined;
+    id?: string | undefined;
+};
+export declare function DataTable(props: DataTableProps): any;
+export declare namespace DataTable {
+    var _HCReact: {
+        type: string;
+        HCOption: string;
+        childOption: string;
+        defaultOptions: any;
+        isArrayType: boolean;
+    };
+}
+export default DataTable;