+++ options/DataTable.js
@@ -1,0 +1,22 @@
+/**
+ * React integration.
+ * Copyright (c) 2026, Highsoft
+ *
+ * A valid license is required for using this software.
+ * See highcharts.com/license
+ *
+ * Built for Highcharts v13.0.0.
+ * Build stamp: 2026-06-30
+ *
+ */
+export function DataTable(props) {
+    return null;
+}
+DataTable._HCReact = {
+    type: "HC_Option",
+    HCOption: "dataTable",
+    childOption: "",
+    defaultOptions: undefined,
+    isArrayType: true,
+};
+export default DataTable;