+++ options/index.d.ts
@@ -5,8 +5,8 @@
  * A valid license is required for using this software.
  * See highcharts.com/license
  *
- * Built for Highcharts v12.6.0.
- * Build stamp: 2026-06-10
+ * Built for Highcharts v13.0.0.
+ * Build stamp: 2026-06-30
  *
  */
 export { default as Title } from "./Title.js";
@@ -17,3 +17,5 @@ export { default as Legend } from "./Legend.js";
 export { default as XAxis } from "./XAxis.js";
 export { default as YAxis } from "./YAxis.js";
 export { default as PlotOptions } from "./PlotOptions.js";
+export { default as DataTable } from "./DataTable.js";
+export { default as Palette } from "./Palette.js";