+++ series/index.js
@@ -5,15 +5,14 @@
  * A valid license is required for using this software.
  * See highcharts.com/license
  *
- * Built for Highcharts v12.6.0.
- * Build stamp: 2026-06-10
+ * Built for Highcharts v13.0.0.
+ * Build stamp: 2026-06-30
  *
  */
 export { default as Area } from "./Area.js";
 export { default as AreaSpline } from "./AreaSpline.js";
 export { default as Bar } from "./Bar.js";
 export { default as Column } from "./Column.js";
-export { default as Heatmap } from "./Heatmap.js";
 export { default as Line } from "./Line.js";
 export { default as Pie } from "./Pie.js";
 export { default as Scatter } from "./Scatter.js";